Added ID3 tag selection.

+++ Form1.cs
@@ -27,6 +27,8 @@ namespace WavToMP3
                 textBoxArtistName.Text = new DirectoryInfo(directoryName).Name;
                 Application.DoEvents();
                 DialogResult result = MessageBox.Show("Ready to start the conversion?", "Confirmation", MessageBoxButtons.YesNo);
+                bool success;
+
                 if (result == System.Windows.Forms.DialogResult.Yes)
                 {
                     var files = System.IO.Directory.EnumerateFiles(directoryName, "*.wav");
@@ -35,8 +37,13 @@ namespace WavToMP3
                     {
                         textBoxProgress.AppendText("Processing " + fileName + "\n");
                         Application.DoEvents();
-                        WavToMP3(fileName, fileName.Replace(".wav", ".mp3"), Int32.Parse(comboBoxBitRate.Text), textBoxArtistName.Text, textBoxAlbumName.Text, checkBoxID3Tags.Checked);
-                        if (checkBoxDeleteSourceFiles.Checked)
+                        success = WavToMP3(fileName, fileName.Replace(".wav", ".mp3"), Int32.Parse(comboBoxBitRate.Text), textBoxArtistName.Text, textBoxAlbumName.Text, checkBoxID3Tags.Checked, listBoxGenre.SelectedIndex.ToString());
+                        if (!success)
+                        {
+                            textBoxProgress.AppendText("Couldn't convert " + fileName + "\n");
+                        }
+
+                        if (success && checkBoxDeleteSourceFiles.Checked)
                         {
                             textBoxProgress.AppendText("Deleting " + fileName + "\n");
                             System.IO.File.Delete(fileName);
@@ -60,30 +67,49 @@ namespace WavToMP3
         /// <param name="artist">Optional artist name</param>
         /// <param name="album">Optional album name</param>
         /// <param name="setID3Tags">Set ID3 tags</param>
-        public static void WavToMP3(string waveFileName, string mp3FileName, int bitRate = 128, string artist = null, string album = null, bool setID3Tags = false)
+        public static bool WavToMP3(string waveFileName, string mp3FileName, int bitRate = 128, string artist = null, string album = null, bool setID3Tags = false, string genre = "148")
         {
-            ID3TagData tags = new ID3TagData();
-            if (setID3Tags)
+            bool result = true;
+            try
             {
-                if (!String.IsNullOrEmpty(artist))
+                ID3TagData tags = new ID3TagData();
+                if (setID3Tags)
                 {
-                    tags.Artist = artist;
-                    tags.Album = album;
-                    tags.Genre = String.Empty;
+                    if (!String.IsNullOrEmpty(artist))
+                    {
+                        tags.Artist = artist;
+                        tags.Album = album;
+                        tags.Genre = genre;
+                    }
                 }
-            }
 
-            using (var reader = new WaveFileReader(waveFileName))
-            using (var writer = new LameMP3FileWriter(mp3FileName, reader.WaveFormat, bitRate, tags))
-                reader.CopyTo(writer);
+                using (var reader = new WaveFileReader(waveFileName))
+                using (var writer = new LameMP3FileWriter(mp3FileName, reader.WaveFormat, bitRate, tags))
+                    reader.CopyTo(writer);
+
+            }
+            catch (Exception)
+            {
+                result = false;
+            }
+            return result;
         }
 
-        // Convert MP3 file to WAV using NAudio classes only
-        public static void MP3ToWave(string mp3FileName, string waveFileName)
+        /// <summary>
+        /// Convert MP3 file to WAV using NAudio classes only
+        /// </summary>
+        /// <param name="mp3FileName">MP3 filename</param>
+        /// <param name="waveFileName">WAV filename</param>
+        public static void MP3ToWav(string mp3FileName, string waveFileName)
         {
             using (var reader = new Mp3FileReader(mp3FileName))
             using (var writer = new WaveFileWriter(waveFileName, reader.WaveFormat))
                 reader.CopyTo(writer);
+        }
+
+        private void checkBoxID3Tags_CheckedChanged(object sender, EventArgs e)
+        {
+            textBoxArtistName.Enabled = textBoxAlbumName.Enabled =  listBoxGenre.Enabled =  checkBoxID3Tags.Checked; 
         }
 
     }
